Frontend/requirements.txt
> python==3.10.16


numpy==1.23.4 
pandas==2.2.2
scikit-learn==1.2.2
xgboost==2.1.3 
flask
mysql-connector-python